+++ examples/get-started/pure-js/widgets/app.js
@@ -8,8 +8,6 @@ import {
   CompassWidget,
   ZoomWidget,
   FullscreenWidget,
-  ScreenshotWidget,
-  ResetViewWidget,
   DarkGlassTheme,
   LightGlassTheme
 } from '@deck.gl/widgets';
@@ -79,8 +77,6 @@ new Deck({
   widgets: [
     new ZoomWidget({style: widgetTheme}),
     new CompassWidget({style: widgetTheme}),
-    new FullscreenWidget({style: widgetTheme}),
-    new ScreenshotWidget({style: widgetTheme}),
-    new ResetViewWidget({style: widgetTheme})
+    new FullscreenWidget({style: widgetTheme})
   ]
 });