+++ examples/website/orthographic/app.tsx
@@ -8,7 +8,7 @@ import {createRoot} from 'react-dom/client';
 import {OrthographicView} from '@deck.gl/core';
 import {TextLayer, PathLayer} from '@deck.gl/layers';
 import {SimpleMeshLayer} from '@deck.gl/mesh-layers';
-import DeckGL from '@deck.gl/react';
+import {DeckGL} from '@deck.gl/react';
 import {Matrix4} from '@math.gl/core';
 import {scaleLinear} from 'd3-scale';
 import {minIndex, maxIndex} from 'd3-array';