+++ modules/core/src/lib/constants.ts
@@ -86,12 +86,13 @@ export const UNIT = {
   pixels: 2
 } as const;
 
-export const EVENT_HANDLERS: {[eventName: string]: string} = {
+export const EVENT_HANDLERS = {
   click: 'onClick',
+  dblclick: 'onClick',
   panstart: 'onDragStart',
   panmove: 'onDrag',
   panend: 'onDragEnd'
-} as const;
+} as const satisfies {[eventName: string]: string};
 
 export const RECOGNIZERS = {
   multipan: [Pan, {threshold: 10, direction: InputDirection.Vertical, pointers: 2}],
@@ -100,6 +101,7 @@ export const RECOGNIZERS = {
   dblclick: [Tap, {event: 'dblclick', taps: 2}],
   click: [Tap, {event: 'click'}, null, ['dblclick']]
 } as const;
+
 export type RecognizerOptions = {
   pinch?: Omit<PinchRecognizerOptions, 'event' | 'enable'>;
   multipan?: Omit<PanRecognizerOptions, 'event' | 'enable'>;