+++ test/modules/core/lib/widget-manager.spec.ts
@@ -6,26 +6,42 @@
 import test from 'tape-promise/tape';
 
 import {WidgetManager} from '@deck.gl/core/lib/widget-manager';
-import {WebMercatorViewport} from '@deck.gl/core';
+import {Widget, WebMercatorViewport, type WidgetProps, type WidgetPlacement} from '@deck.gl/core';
 
-class TestWidget {
-  constructor(props) {
-    this.id = props.id;
-    this.viewId = props.viewId || null;
-    this.placement = props.placement || 'top-left';
-    this.props = props;
+type TestWidgetProps = WidgetProps & {
+  placement?: WidgetPlacement;
+  viewId?: string;
+  version?: number;
+};
+class TestWidget extends Widget<TestWidgetProps> {
+  static defaultProps: Required<TestWidgetProps> = {
+    ...Widget.defaultProps,
+    id: 'test-widget',
+    placement: 'top-left',
+    version: 1,
+    viewId: 'default-view'
+  };
+
+  placement: WidgetPlacement = 'top-left';
+  className = 'deck-test-widget';
+  isVisible = false;
+
+  constructor(props: TestWidgetProps = {}) {
+    super(props, TestWidget.defaultProps);
+    this.viewId = props.viewId ?? this.viewId;
+    this.placement = props.placement ?? 'top-left';
   }
 
+  onRenderHTML(rootElement: HTMLElement): void {}
+
   setProps(props) {
-    Object.assign(this.props, props);
+    this.viewId = props.viewId ?? this.viewId;
+    this.placement = props.placement ?? this.placement;
+    super.setProps(props);
   }
 
   onAdd() {
     this.isVisible = true;
-
-    const el = document.createElement('div');
-    el.id = this.id;
-    return el;
   }
 
   onRemove() {
@@ -50,7 +66,7 @@ test('WidgetManager#setProps', t => {
   t.is(widgetManager.getWidgets().length, 1, 'widget is added');
   t.ok(widgetA.isVisible, 'widget.onAdd is called');
   t.ok(
-    widgetManager.containers['__root'].contains(widgetA._element),
+    widgetManager.containers['__root'].contains(widgetA.rootElement),
     'widget UI is added to the container'
   );
   t.is(container.childElementCount, 1, 'widget container is added');
@@ -63,18 +79,18 @@ test('WidgetManager#setProps', t => {
   t.is(widgetManager.getWidgets().length, 2, 'widget is added');
   t.ok(widgetB.isVisible, 'widget.onAdd is called');
   t.ok(
-    widgetManager.containers['map'].contains(widgetB._element),
+    widgetManager.containers['map'].contains(widgetB.rootElement),
     'widget UI is added to the container'
   );
   t.is(container.childElementCount, 2, 'widget container is added');
 
-  const elementA = widgetA._element;
+  const elementA = widgetA.rootElement;
   // Only B
   widgetManager.setProps({
     widgets: [widgetB]
   });
   t.is(widgetManager.getWidgets().length, 1, 'widget is removed');
-  t.notOk(widgetA._element, 'widget context is cleared');
+  t.notOk(widgetA.rootElement, 'widget context is cleared');
   t.notOk(widgetA.isVisible, 'widget.onRemove is called');
   t.notOk(
     widgetManager.containers['__root'].contains(elementA),
@@ -340,17 +356,18 @@ test('WidgetManager#onHover, onEvent#without viewId', t => {
   widget.onHover = () => onHoverCalledCount++;
   widget.onClick = () => onClickCalledCount++;
 
-  // Trigger onHover event
+  // Given the pickedInfo, test that widgetManager does forward events to test widget
+  // Trigger hover event leading to onHover callback
   widgetManager.onHover(pickedInfo, {});
-  // Trigger onClick event
+  // Trigger click event leading to onClick callback
   widgetManager.onEvent(pickedInfo, {type: 'click'});
-  // Trigger onDragStart event
+  // Trigger panstart event leading to onDragStart callback
   widgetManager.onEvent(pickedInfo, {type: 'panstart'});
-  // Event not defined
+  // Trigger dblclick event leading to onClick callback
   widgetManager.onEvent(pickedInfo, {type: 'dblclick'});
 
   t.is(onHoverCalledCount, 1, 'widget.onHover is called');
-  t.is(onClickCalledCount, 1, 'widget.onClick is called');
+  t.is(onClickCalledCount, 2, 'widget.onClick is called');
 
   widgetManager.finalize();
   t.end();
@@ -373,30 +390,32 @@ test('WidgetManager#onHover, onEvent#viewId', t => {
   widget.onHover = () => onHoverCalledCount++;
   widget.onClick = () => onClickCalledCount++;
 
-  // Trigger onHover event
+  // Given the pickedInfo, test that widgetManager does forward events to test widget
+  // Trigger hover event leading to onHover callback
   widgetManager.onHover(pickedInfo, {});
-  // Trigger onClick event
+  // Trigger click event leading to onClick callback
   widgetManager.onEvent(pickedInfo, {type: 'click'});
-  // Trigger onDragStart event
+  // Trigger panstart event leading to onDragStart callback
   widgetManager.onEvent(pickedInfo, {type: 'panstart'});
-  // Event not defined
+  // Trigger dblclick event leading to onClick callback
   widgetManager.onEvent(pickedInfo, {type: 'dblclick'});
 
   t.is(onHoverCalledCount, 1, 'widget.onHover is called');
-  t.is(onClickCalledCount, 1, 'widget.onClick is called');
+  t.is(onClickCalledCount, 2, 'widget.onClick is called');
 
   pickedInfo = {
     viewport: new WebMercatorViewport({id: 'minimap'}),
     index: 0
   };
 
-  // Trigger onHover event
+  // Given the updated pickedInfo, test that widgetManager does *not* forward events to test widget
+  // Trigger hover event not leading to onHover callback
   widgetManager.onHover(pickedInfo, {});
-  // Trigger onClick event
+  // Trigger click event not leading to onClick callback
   widgetManager.onEvent(pickedInfo, {type: 'click'});
 
   t.is(onHoverCalledCount, 1, 'widget.onHover is not called');
-  t.is(onClickCalledCount, 1, 'widget.onClick is not called');
+  t.is(onClickCalledCount, 2, 'widget.onClick is not called');
 
   widgetManager.finalize();
   t.end();