+++ test/modules/geo-layers/index.ts
@@ -5,6 +5,7 @@
 import test from 'tape-promise/tape';
 
 import {
+  A5Layer,
   GreatCircleLayer,
   H3HexagonLayer,
   H3ClusterLayer,
@@ -18,6 +19,7 @@ import {
 } from '@deck.gl/geo-layers';
 
 test('Top-level imports', t => {
+  t.ok(A5Layer, 'A5Layer symbol imported');
   t.ok(GreatCircleLayer, 'GreatCircleLayer symbol imported');
   t.ok(QuadkeyLayer, 'QuadkeyLayer symbol imported');
   t.ok(S2Layer, 'S2Layer symbol imported');
@@ -31,6 +33,7 @@ test('Top-level imports', t => {
   t.end();
 });
 
+import './a5-layer.spec';
 import './tile-layer';
 import './wms-layer.spec';
 import './quadkey-layer.spec';